# #2785 - Faster SIH of a linear map of a hyperrectangle