Closes #2785.

```julia
julia> H = rand(Hyperrectangle, dim=1000);


julia> M = sprand(1000, 1000, 0.01);

julia> @time symmetric_interval_hull(M*H);  # master
 29.072374 seconds (2.01 k allocations: 15.534 MiB, 0.04% gc time)

julia> @time symmetric_interval_hull(M*H);  # this branch
  0.000179 seconds (11 allocations: 195.625 KiB)


julia> M = rand(1000, 1000);

julia> @time symmetric_interval_hull(M*H);  # master
  2.494493 seconds (2.01 k allocations: 15.534 MiB)

julia> @time symmetric_interval_hull(M*H);  # this branch
  0.009323 seconds (8 allocations: 7.661 MiB)
```